fix: 스프린트 미션2 피드백 적용 완료
