+++ css/auth.css
@@ -1,38 +1,38 @@
-.login-page {
+.auth-page {
   padding-block: 230px;
 }
-.login-page .inner03 {
+.auth-page .inner03 {
   display: flex;
   flex-direction: column;
   gap: 24px;
 }
-.login-logo {
+.auth-logo {
   display: block;
   margin-bottom: 40px;
-  width: 396px;
+  width: 386px;
   margin-inline: auto;
 }
-.login-logo img {
+.auth-logo img {
   width: 100%;
 }
-.login-form {
+.auth-form {
   display: flex;
   flex-direction: column;
   gap: 24px;
 }
-.login-form li {
+.auth-form li {
   display: flex;
   flex-direction: column;
   gap: 16px;
   position: relative;
 }
-.login-form li label {
+.auth-form li label {
   color: var(--gray800);
   font-size: 18px;
   font-weight: 700;
 }
-.login-form li input[type='text'],
-.login-form li input[type='password'] {
+.auth-form li input[type='text'],
+.auth-form li input[type='password'] {
   height: 56px;
   background-color: #f3f4f6;
   border-radius: 12px;
@@ -40,7 +40,7 @@
   font-size: 16px;
   font-weight: 400;
 }
-.login-btn {
+.auth-btn {
   display: flex;
   background-color: var(--blue);
   color: #fff;
@@ -51,32 +51,33 @@
   border-radius: 2em;
   height: 56px;
   width: 100%;
+  margin-top: 24px;
 }
-.login-sns {
+.auth-sns {
   display: flex;
   align-items: center;
   padding: 16px 24px;
   background-color: #e6f2ff;
   justify-content: space-between;
 }
-.login-sns p {
+.auth-sns p {
   color: var(--gray800);
   font-size: 16px;
   font-weight: 500;
 }
-.login-sns ul {
+.auth-sns ul {
   display: flex;
   align-items: center;
   gap: 16px;
 }
-.login-sns ul li a {
+.auth-sns ul li a {
   width: 42px;
   display: block;
 }
-.login-sns ul li a img{
+.auth-sns ul li a img{
   width: 100%;
 }
-.login-signup {
+.auth-signup {
   color: var(--gray800);
   font-size: 14px;
   font-weight: 500;
@@ -85,7 +86,7 @@
   align-items: center;
   gap: 4px;
 }
-.login-signup a {
+.auth-signup a {
   color: var(--blue);
   text-decoration-line: underline;
   text-underline-offset: 2px;
@@ -98,27 +99,32 @@
   right: 24px;
   bottom: 28px;
   transform: translateY(50%);
+  width: 20px;
+  height: 18px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
 }
 
 @media all and (max-width: 640px) {
-  .login-page {
+  .auth-page {
     padding-block: 35.94vw;
   }
-  .login-logo {
+  .auth-logo {
     margin-bottom: 8.59vw;
     width: 61.88vw;
   }
-  .login-page .inner03 {
+  .auth-page .inner03 {
     gap: 5vw;
   }
-  .login-form li {
+  .auth-form li {
     gap: 3.75vw;
   }
-  .login-form li label {
+  .auth-form li label {
     font-size: 4.06vw;
   }
-  .login-form li input[type='text'],
-  .login-form li input[type='password'] {
+  .auth-form li input[type='text'],
+  .auth-form li input[type='password'] {
     height: 12.5vw;
     padding-inline: 3.44vw;
     font-size: 3.75vw;
@@ -132,23 +138,23 @@
   .toggle-password img {
     max-width: 4.69vw;
   }
-  .login-sns {
+  .auth-sns {
     padding: 2.5vw 3.75vw;
   }
-  .login-btn {
+  .auth-btn {
     font-size: 4.38vw;
     height: 12.5vw;
   }
-  .login-sns p {
+  .auth-sns p {
     font-size: 3.75vw;
   }
-  .login-sns ul {
+  .auth-sns ul {
     gap: 3.13vw;
   }
-  .login-sns ul li a {
+  .auth-sns ul li a {
     width: 9.38vw;
   }
-  .login-signup {
+  .auth-signup {
     font-size: 3.44vw;
   }
 }
