+++ css/main.css
@@ -31,19 +31,19 @@
   line-height: 1.3;
   margin-top: 24px;
 }
-.main-common figure {
+.main-common .img {
   width: 588px;
   overflow: hidden;
   aspect-ratio: 1.324 / 1;
 }
-.main-common figure img {
+.main-common .img img {
   width: 100%;
   height: 100%;
   object-fit: cover;
   object-position: center;
   transition: transform 0.8s;
 }
-.main-common a:hover figure img {
+.main-common a:hover .img img {
   transform: scale(1.1);
 }
